Shuffle the PropTestClient init location
If we make prop tests only depend on a custom client abstraction, it makes the code depend less explicitly on existing in its own project with the actual OpenSearchClient. Would help if we want to later integrate tests like these in the rest of our CI ecosystem. Probably need to do more work to abstract away other client uses, though.

Signed-off-by: Simeon Widdis <sawiddis@amazon.com>